requirements.txt
bert_score==0.3.13
bioregistry==0.11.35
colorcet==3.1.0
datasets==2.19.0
elasticsearch==9.0.0
faiss_cpu==1.9.0.post1
falcon==4.0.2
filelock==3.18.0
FlagEmbedding==1.3.3
FlagEmbedding==1.3.4
gliner==0.2.16
Jinja2==3.1.6
joblib==1.4.2
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.52
langchain_openai==0.3.13
local_gemma==0.2.0
matplotlib==3.10.1
nltk==3.9.1
numpy==2.2.4
openpyxl==3.1.5
optimum==1.23.3
pandas==2.2.3
peft==0.14.0
pycountry==24.6.1
pydantic==2.11.3
pymongo==4.12.0
python-dotenv==1.1.0
python_dateutil==2.9.0.post0
rapidfuzz==3.13.0
Requests==2.32.3
rich==14.0.0
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
seqeval==1.2.2
spacy==3.7.5
statsmodels==0.14.4
tenacity==9.0.0
torch==2.5.1+cu121
tqdm==4.67.1
transformers==4.49.0